Feat: domain-base routing (#99)
