+++ packages/qwik-speak/src/index.ts
@@ -6,25 +6,21 @@ export type {
   SpeakConfig,
   SpeakState,
   LoadTranslationFn,
-  RewriteRouteOption
+  RewriteRouteOption,
+  DomainBasedRoutingOption
 } from './types';
 export type { QwikSpeakProps, QwikSpeakMockProps } from './use-qwik-speak';
 export type { SpeakProps } from './use-speak';
-export type { LocalizePathFn } from './localize-path';
-export type { TranslatePathFn } from './translate-path';
 export type { InlinePluralFn } from './inline-plural';
 export type { InlineTranslateFn } from './inline-translate';
 export type { FormatDateFn } from './use-format-date';
 export type { FormatNumberFn } from './use-format-number';
 export type { RelativeTimeFn } from './use-relative-time';
 export type { DisplayNameFn } from './use-display-name';
+export type { LocalizePathFn, TranslatePathFn } from './routing';
 // Inline functions
 export { inlineTranslate } from './inline-translate';
 export { inlinePlural } from './inline-plural';
-// Functions
-export { localizePath } from './localize-path';
-export { translatePath } from './translate-path';
-export { validateLocale } from './validate-locale';
 // Use functions
 export { useQwikSpeak } from './use-qwik-speak';
 export { useSpeak } from './use-speak';
@@ -37,5 +33,13 @@ export {
   useSpeakLocale,
   useSpeakConfig,
 } from './use-functions';
+// Routing
+export {
+  localizePath,
+  translatePath,
+  validateLocale,
+  extractFromDomain,
+  extractFromUrl
+} from './routing';
 // Testing
 export { QwikSpeakMockProvider } from './use-qwik-speak';
