+++ packages/qwik-speak/src/types.ts
@@ -30,6 +30,14 @@ export interface SpeakLocale {
    * Text direction
    */
   dir?: 'ltr' | 'rtl' | 'auto';
+  /**
+   * In domain-based routing, set the default domain for the locale
+   */
+  domain?: string;
+  /**
+   * In domain-based routing, set another domain for the locale
+   */
+  withDomain?: string;
 }
 
 /**
@@ -59,6 +67,13 @@ export interface RewriteRouteOption {
    * Key value pairs: folder name - translated value
    */
   paths: Record<string, string>;
+}
+
+export interface DomainBasedRoutingOption {
+  /**
+   * Always use the lang prefix in domain-based routing, or as needed
+   */
+  prefix: 'always' | 'as-needed'
 }
 
 export interface SpeakConfig {
@@ -91,6 +106,10 @@ export interface SpeakConfig {
    * Rewrite routes as specified in Vite config for qwikCity
    */
   rewriteRoutes?: RewriteRouteOption[];
+  /**
+   * Domain-based routing options
+   */
+  domainBasedRouting?: DomainBasedRoutingOption
 }
 
 export interface SpeakState {
